+++ tests/TestCase/TestSuite/TestCaseTest.php
@@ -14,7 +14,12 @@ declare(strict_types=1);
  */
 namespace MeTools\Test\TestCase\TestSuite;
 
-use App\Test\TestCase\View\AppViewTest;
+use AnotherTestPlugin\Test\TestCase\Controller\MyExampleControllerTest;
+use App\Test\TestCase\BadTestClass;
+use App\Test\TestCase\Controller\PagesControllerTest;
+use App\Test\TestCase\Model\Table\PostsTableTest;
+use App\Test\TestCase\Model\Validation\PostValidatorTest;
+use MeTools\Test\TestCase\View\Helper\HtmlHelperTest;
 use MeTools\TestSuite\TestCase;
 use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\ExpectationFailedException;
@@ -26,7 +31,7 @@ use Tools\Filesystem;
 class TestCaseTest extends TestCase
 {
     /**
-     * @var \MeTools\TestSuite\TestCase&\PHPUnit\Framework\MockObject\MockObject
+     * @var \MeTools\TestSuite\TestCase
      */
     protected TestCase $TestCase;
 
@@ -35,24 +40,9 @@ class TestCaseTest extends TestCase
      */
     public function setUp(): void
     {
-        $this->TestCase ??= $this->getMockForAbstractClass(TestCase::class);
-    }
-
-    /**
-     * @test
-     * @uses \MeTools\TestSuite\TestCase::__get()
-     */
-    public function testGetMagicMethod(): void
-    {
-        $AppViewTest = new AppViewTest();
-        $this->assertSame('App\View\AppView', $AppViewTest->originClassName);
-        $this->assertSame('App', $AppViewTest->alias);
-
-        //With a no existing property
-        $this->expectException(AssertionFailedError::class);
-        $this->expectExceptionMessage('Property `noExistingProperty` does not exist');
-        /** @noinspection PhpUndefinedFieldInspection */
-        $AppViewTest->noExistingProperty;
+        $this->TestCase ??= new class ('MyTest') extends TestCase
+        {
+        };
     }
 
     /**
@@ -68,7 +58,7 @@ class TestCaseTest extends TestCase
         }
 
         $this->expectException(ExpectationFailedException::class);
-        $this->expectExceptionMessage('Failed asserting that \'' . $string . '\' contains "bad word".');
+        $this->expectExceptionMessageMatches('/^Failed asserting that \'cat dog bird\'( \[ASCII\]\(length: 12\))? contains "bad word"( \[ASCII\]\(length: 8\))?\.$/');
         $this->TestCase->assertLogContains('bad word', LOGS . 'debug.log');
     }
 
@@ -81,5 +71,55 @@ class TestCaseTest extends TestCase
         $this->expectException(AssertionFailedError::class);
         $this->expectExceptionMessage('Failed asserting that file "' . LOGS . 'noExisting.log" exists');
         $this->assertLogContains('content', LOGS . 'noExisting.log');
+    }
+
+    /**
+     * @test
+     * @uses \MeTools\TestSuite\TestCase::getAlias()
+     */
+    public function testGetAlias(): void
+    {
+        foreach ([
+            'Pages' => PagesControllerTest::class,
+            'Html' => HtmlHelperTest::class,
+            'Posts' => PostsTableTest::class,
+            'Post' => PostValidatorTest::class,
+         ] as $expectedAlias => $testClassName) {
+            /** @var \MeTools\TestSuite\TestCase $TestClass */
+            $TestClass = new $testClassName($expectedAlias . 'Test');
+            $this->assertSame($expectedAlias, $TestClass->getAlias());
+        }
+
+        $this->expectException(AssertionFailedError::class);
+        $this->expectExceptionMessage('Unable to get the alias for `App\Test\TestCase\BadTestClass`');
+        $TestClass = new BadTestClass('BadTest');
+        $TestClass->getAlias();
+    }
+
+    /**
+     * @test
+     * @uses \MeTools\TestSuite\TestCase::getOriginClassName()
+     */
+    public function testGetOriginClassName(): void
+    {
+        $TestClass = new TestCaseTest('MyTest');
+        $this->assertSame(TestCase::class, $TestClass->getOriginClassName());
+
+        $this->expectException(AssertionFailedError::class);
+        $this->expectExceptionMessage('Unable to determine the origin class for `App\Test\TestCase\BadTestClass`');
+        $TestClass = new BadTestClass('BadTest');
+        $TestClass->getOriginClassName();
+    }
+
+    /**
+     * @test
+     * @uses \MeTools\TestSuite\TestCase::getOriginClassName()
+     */
+    public function testGetOriginClassNameWithNoExistingClass(): void
+    {
+        $this->expectException(AssertionFailedError::class);
+        $this->expectExceptionMessage('Class `AnotherTestPlugin\Controller\MyExampleController` does not exist');
+        $TestClass = new MyExampleControllerTest('MyExampleControllerTest');
+        $TestClass->getOriginClassName();
     }
 }